Refactor new_graph QMenu
