+++ source/view/components/new_graph_dialog.py
